'File Provider Storage' --> 'tmp'
This is to match changes made in [Working Copy](https://workingcopyapp.com) but it only solves part of the problem.  Working Copy now shares content with other apps using the `tmp` directory instead of the `File Provider Storage`.  This problem is fixed.  Working Copy now zips up folders into a single zip file before sharing them.  This code just drops that zip file into the `from Working Copy` folder and leaves the unzipping for the user to do.  Look for an update soon to automate the unzipping.